Price Section

+++ database/migrations/2022_05_24_093152_create_property_images_table.php
@@ -15,7 +15,7 @@ class CreatePropertyImagesTable extends Migration
     {
         Schema::create('property_images', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('property_id');
+            $table->unsignedBigInteger('property_id');
             $table->string('images');
 
             $table->timestamps();
